requirements.txt
Django==1.11.4
Pillow==4.2.1
psycopg2==2.7.3.1
pycurl==7.43.0
pygobject==3.20.0
python-apt==1.1.0b1
